requirements.txt
beautifulsoup4==4.12.3
googlesearch_python==1.2.5
langchain==0.3.7
langchain-core==0.3.15
langchain-openai==0.2.6
langgraph==0.2.45
openai==1.54.1
requests==2.32.3
tqdm==4.66.6
